[RENAME] NewGift -> Gift로 폴더 명 수정
