+++ Favor/Favor/Sources/Scenes/Gift/Views/NewGiftFriendFooterView.swift
@@ -35,8 +35,18 @@ final class NewGiftFriendFooterView: UICollectionReusableView, Reusable {
     return iv
   }()
   
+  private lazy var tapGesture: UITapGestureRecognizer = {
+    let tg = UITapGestureRecognizer()
+    tg.addTarget(self, action: #selector(self.didTapFooterView))
+    return tg
+  }()
+  
   private let divider = FavorDivider()
-
+  
+  // MARK: - Properties
+  
+  var tapCompletion: (() -> Void)?
+  
   // MARK: - Initializer
 
   override init(frame: CGRect) {
@@ -49,11 +59,19 @@ final class NewGiftFriendFooterView: UICollectionReusableView, Reusable {
   required init?(coder: NSCoder) {
     fatalError("init(coder:) has not been implemented")
   }
+  
+  // MARK: - Selectors
+  
+  @objc
+  private func didTapFooterView() {
+    self.tapCompletion?()
+  }
 }
 
 extension NewGiftFriendFooterView: BaseView {
   func setupStyles() {
     self.backgroundColor = .favorColor(.white)
+    self.addGestureRecognizer(self.tapGesture)
   }
   
   func setupLayouts() {
