--- Favor/Favor/Sources/Scenes/NewGift/NewGiftPhotoSection.swift
@@ -1,19 +1,0 @@
-//
-//  PickedPictureSection.swift
-//  Favor
-//
-//  Created by 김응철 on 2023/02/08.
-//
-
-import UIKit
-
-import RxDataSources
-
-struct NewGiftPhotoSection {
-  typealias NewGiftPhotoSectionModel = SectionModel<Int, NewGiftSectionItem>
-  
-  enum NewGiftSectionItem: Equatable {
-    case empty
-    case photo(UIImage)
-  }
-}
