--- Favor/Favor/Sources/Scenes/NewGift/ViewControllers/ShareVC.swift
@@ -1,14 +1,0 @@
-//
-//  ShareVC.swift
-//  Favor
-//
-//  Created by 이창준 on 2023/01/11.
-//
-
-import UIKit
-
-import FavorKit
-
-final class ShareViewController: BaseViewController {
-  
-}
